+++ screens/Home.js
@@ -8,7 +8,7 @@ import Column from '../components/Column';
 import { Card, Divider, Layout, List, ListItem, Text } from '@ui-kitten/components';
 import moment from 'moment';
 import Duet from '../components/Duet';
-import SectionHeading from '../components/SectionHeading';
+import Mark from '../components/Mark';
 import Culture from '../components/Culture';
 import Shelf from '../components/Shelf'; // Could probably have an index file with all the sections and import all of them on one line.
 import Featured from "../components/Featured";
@@ -28,17 +28,17 @@ export default function Home(props) {
       <Layout style={styles.container}>
         <ScrollView>
           <Featured articles={dummyData.slice(40, 45)} />
-          <SectionHeading title={Sections.NEWS.name} onPress={() => console.log("navigate to next screen")}/>
+          <Mark title={Sections.NEWS.name} onPress={() => console.log("navigate to next screen")}/>
           <Duet articles={dummyData.slice(0, 10)} />
           <Divider marginTop={8} />
-          <SectionHeading title={Sections.OPINIONS.name} onPress={() => console.log("navigate to next screen")}/>
+          <Mark title={Sections.OPINIONS.name} onPress={() => console.log("navigate to next screen")}/>
           <Shelf articles={dummyData.slice(10, 20)} />
-          <SectionHeading title={Sections.SPORTS.name} onPress={() => console.log("navigate to next screen")}/>
+          <Mark title={Sections.SPORTS.name} onPress={() => console.log("navigate to next screen")}/>
           <Duet articles={dummyData.slice(20, 30)} />
           <Divider marginTop={8} />
           <Culture theGrind={dummyData.slice(30, 32)} artsAndLife={dummyData.slice(32,34)} />
           <Divider />
-          <SectionHeading alternate title={Sections.HUMOR.name} onPress={() => console.log("navigate to next screen")}/>
+          <Mark alternate title={Sections.HUMOR.name} onPress={() => console.log("navigate to next screen")}/>
           <Shelf alternate articles={dummyData.slice(33, 40)} />
           {/* Infinite scroll/wildcard will go here, with cell similar to the ones in `Culture` component. */}
         </ScrollView>
